Appointly - services, inMemoryRepositories and services tests

+++ src/main/java/com/ania/appointly/application/service/UserService.java
@@ -3,24 +3,29 @@ import com.ania.appointly.application.usecase.user.CreateUserUseCase;
 import com.ania.appointly.application.usecase.user.DeleteUserUseCase;
 import com.ania.appointly.application.usecase.user.ReadUserUseCase;
 import com.ania.appointly.application.usecase.user.UpdateUserUseCase;
+import com.ania.appointly.domain.exeptions.UserValidationException;
 import com.ania.appointly.domain.model.Role;
 import com.ania.appointly.domain.model.User;
 import com.ania.appointly.domain.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+@Service
+@RequiredArgsConstructor
 public class UserService implements
         CreateUserUseCase,  ReadUserUseCase,
         UpdateUserUseCase, DeleteUserUseCase {
-    private final UserRepository userRepository;
 
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+    private final UserRepository userRepository;
 
     @Override
     public User createUser(User user) {
+        if(userRepository.existsById(user.getId())){
+            throw new UserValidationException("User with this ID already exists.");
+        }
         return userRepository.save(user);
     }
 
@@ -51,11 +56,17 @@ public class UserService implements
 
     @Override
     public User updateUser(User user) {
+        if(!userRepository.existsById(user.getId())){
+            throw new UserValidationException("User with this ID does not exist.");
+        }
         return userRepository.save(user);
     }
 
     @Override
     public void deleteUser(UUID id) {
+        if(!userRepository.existsById(id)){
+            throw new UserValidationException("User with this ID does not exist.");
+        }
         userRepository.deleteById(id);
     }
 }
